Renamed database/queries for consistent naming conventions, introduced typecast from BIT to boolean for mysql & node.js. Tested
