+++ database/queries/index.js
@@ -1,3 +1,3 @@
-module.exports.getUserByGoogleID = require('./getUserByGoogleID');
-module.exports.getUserByID = require('./getUserByID');
-module.exports.insertUser = require('./insertUser');
+module.exports.getUserByGoogleID = require('./get-user-by-google-id');
+module.exports.getUserByID = require('./get-user-by-id');
+module.exports.insertUser = require('./insert-user');
